+++ packages/core/src/aminomessages/cosmos/bank/authorizations.ts
@@ -3,6 +3,13 @@ import { SendAuthorization } from "cosmjs-types/cosmos/bank/v1beta1/authz";
 import { Any } from "cosmjs-types/google/protobuf/any";
 import { AminoSendAuthorization } from "./messages";
 
+export function sendAuthorizationToAny(authorization: SendAuthorization): Any {
+  return Any.fromPartial({
+    typeUrl: "/cosmos.bank.v1beta1.SendAuthorization",
+    value: SendAuthorization.encode(authorization).finish(),
+  });
+}
+
 export function createBankAuthorizationConverters(): AminoConverters {
   return {
     "/cosmos.bank.v1beta1.SendAuthorization": {
@@ -17,12 +24,14 @@ export function createBankAuthorizationConverters(): AminoConverters {
       },
       fromAmino: (
         authorization: AminoSendAuthorization["value"]
-      ): Any["value"] =>
-        SendAuthorization.encode(
+      ): Any["value"] => {
+        const any = sendAuthorizationToAny(
           SendAuthorization.fromPartial({
             spendLimit: authorization.spend_limit,
           })
-        ).finish(),
+        );
+        return any.value;
+      },
     },
   };
 }